Move command from script.js
